Move get_canonical_feature_name
