+++ rust/optify/src/builder/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod builder_impl;
 pub(crate) mod builder_options;
 pub(crate) mod builder_trait;
+mod get_canonical_feature_name;
 pub(crate) mod loading_result;
 pub(crate) mod watcher_builder;
